refactor: move Maven discovery logic from workspace.js to java_maven.js
Address review feedback: provider-specific discovery functions don't belong
in workspace.js. Extract traverseForWrapper to tools.js as shared utility,
update base_java.js to delegate to it, and move all Maven workspace discovery
functions (discoverMavenModules, resolveMavenBinary, collectMavenModules,
listMavenModules, parseMavenModuleList) to their provider file.
